requirements.txt
numpy>=1.24.0
scipy>=1.11.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
streamlit>=1.28.0
fpdf2>=2.7.6
pandas>=2.0.0
pytest>=7.4.0
joblib>=1.3.0
